module-4/studio/requirements.txt
langgraph
langchain-core
langchain-community
langchain-groq
python-dotenv
httpx
tavily-python
wikipedia